requirements.txt
numpy
matplotlib
pyqt5
joblib
scipy
theano
numexpr
seaborn
statsmodels
scikit-image
lime
pdpbox
scikit-learn
fastcache
tqdm >= 4.56.2
# https://github.com/NVIDIA/framework-determinism
tensorflow-gpu >= 2.3.0
graphviz
pydot
memory_profiler
alibi
alibi[ray]

# reviewers: there are some scripts that run into issues with this version - we have a fork of sympy that fixes
#  some issues, but have not included it for the sake of anonymity
sympy

saliency

# r interface (requires R installed) - also, please install shapr within R before running SHAPR
rpy2

pygam

mpmath
gmpy
